MPS: always copy /etc/modprobe.d to live/rootfs initramfs
In MPS, the initramfs is being built outside the target system, so it's
always created in non-hostonly mode. However, /etc/modprobe.d on MPS
contains some important files, and needs to be copied to the MPS
initramfs. This is just a workaround.

See https://bugzilla.redhat.com/show_bug.cgi?id=873220#c24 why the
original behavior is correct in generic case.